normalize indexes after delition
